+++ src/templates/frontend/linters/next/eslint/typescript/eslint.config.mjs
@@ -2,6 +2,7 @@ import { FlatCompat } from '@eslint/eslintrc';
 import js from '@eslint/js';
 import typescriptEslint from '@typescript-eslint/eslint-plugin';
 import tsParser from '@typescript-eslint/parser';
+import jsxA11Y from 'eslint-plugin-jsx-a11y';
 import globals from 'globals';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
@@ -22,27 +23,37 @@ export default [
   ),
   {
     plugins: {
+      'jsx-a11y': jsxA11Y,
       '@typescript-eslint': typescriptEslint,
     },
 
     languageOptions: {
       globals: {
-        ...globals.node,
+        ...globals.browser,
       },
 
       parser: tsParser,
       ecmaVersion: 'latest',
       sourceType: 'module',
+
+      parserOptions: {
+        ecmaFeatures: {
+          jsx: true,
+        },
+      },
     },
 
     settings: {
+      react: {
+        version: 'detect',
+      },
+
       'import/parsers': {
         '@typescript-eslint/parser': ['.ts', '.tsx', '.d.ts'],
       },
     },
 
     rules: {
-      'no-useless-constructor': 'off',
       'prettier/prettier': [
         'error',
         {
@@ -52,8 +63,23 @@ export default [
           trailingComma: 'all',
           arrowParens: 'always',
           semi: true,
+          endOfLine: 'auto',
         },
       ],
+
+      'jsx-a11y/alt-text': [
+        'warn',
+        {
+          elements: ['img'],
+          img: ['Image'],
+        },
+      ],
+
+      'jsx-a11y/aria-props': 'warn',
+      'jsx-a11y/aria-proptypes': 'warn',
+      'jsx-a11y/aria-unsupported-elements': 'warn',
+      'jsx-a11y/role-has-required-aria-props': 'warn',
+      'jsx-a11y/role-supports-aria-props': 'warn',
     },
   },
 ];